backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
sqlalchemy==2.0.43
alembic==1.16.5
python-multipart==0.0.20
pytest
pytest-cov
httpx==0.28.1
